requirements.txt
pymongo
lxml
python-dotenv